refactor: moving contexts up to common folder
